requirements_optional.txt
# Core dependencies for authentication and encryption
cryptography>=41.0.0
keyring>=24.0.0
instaloader>=4.9.0

# Build dependencies  
pyinstaller>=6.0.0

# Media processing
ffmpeg-python>=0.2.0

# Additional downloaders
gallery-dl>=1.26.0

# Security and authentication
cryptography>=3.4.8
requests-oauthlib>=1.3.0
keyring>=23.0.0
pycryptodome>=3.15.0

# Error monitoring and analysis
sentry-sdk>=1.5.0

# Session management
requests-cache>=0.9.0

# Optional dependencies for private account access
instaloader>=4.9.0
# TikTok API and scraping
tiktokapipy>=0.2.2
playwright>=1.40.0
# YouTube authentication and enhanced features
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.1
# Enhanced web scraping
selenium>=4.15.0
undetected-chromedriver>=3.5.4
# Session management
requests-oauthlib>=1.3.1
browser-cookie3>=0.19.1
# Additional utilities
fake-useragent>=1.4.0
cloudscraper>=1.2.69